+++ src/kinematics/movementFunctions.cpp
@@ -23,13 +23,13 @@ vec3P doLegLengthCorrection(vec3P givenLegPosition, int givenLegIndex){
     return givenLegPosition;
 }
 
-vec3P currentLegPos(int legId, std::vector<double> servoAnglesInRad, std::vector<double> legActuatorLengths){
+vec3P currentLegPos(int legId, std::vector<double> servoAnglesInRad, std::vector<float> legActuatorLengths){
   vec3P position = forwardKinematics(servoAnglesInRad[3*legId], servoAnglesInRad[(3*legId)+1], servoAnglesInRad[(3*legId)+2], 0.0f, legActuatorLengths[0], legActuatorLengths[1]);
 
   return position;
 }
 
-std::vector<vec3P> currentLegPositions(std::vector<double> servoAnglesInRad, std::vector<double> legLengths){
+std::vector<vec3P> currentLegPositions(std::vector<double> servoAnglesInRad, std::vector<float> legLengths){
   std::vector<vec3P> vectorToReturn(4);
 
   for (int i = 0; i < 4; i++){
